Move SAML authentication from transmart-app to transmart-oauth
