+++ transmart-oauth/build.gradle
@@ -53,6 +53,9 @@ dependencies {
     compile 'org.grails.plugins:spring-security-core:3.1.1'
     compile 'org.grails.plugins:spring-security-oauth2-provider:3.0.0-RC2'
     compile 'org.springframework.security:spring-security-ldap:3.2.3.RELEASE'
-
+    
+    /* for SAML authentication */
+    compile('org.springframework.security.extensions:spring-security-saml2-core:1.0.0.RELEASE')
+    
     compile project(':search-domain')
 }
