+++ transmartApp/grails-app/init/org/transmartproject/app/BootStrap.groovy
@@ -24,13 +24,6 @@ class BootStrap {
 
         SpringSecurityUtils.clientRegisterFilter('concurrentSessionFilter', SecurityFilterPosition.CONCURRENT_SESSION_FILTER)
 
-        if (grailsApplication.config.org.transmart.security.samlEnabled) {
-            SpringSecurityUtils.clientRegisterFilter(
-                    'metadataGeneratorFilter', SecurityFilterPosition.FIRST)
-            SpringSecurityUtils.clientRegisterFilter(
-                    'samlFilter', SecurityFilterPosition.BASIC_AUTH_FILTER)
-        }
-
         if (!grailsApplication.config.org.transmart.configFine.is(true)) {
             logger.error("Something wrong happened parsing the externalized " +
                     "Config.groovy, because we could not find the " +
