+++ transmartApp/src/main/groovy/transmartapp/TransmartAppGrailsPlugin.groovy
@@ -77,22 +77,6 @@ Brief summary/description of the plugin.
             xmlns context: "http://www.springframework.org/schema/context"
             xmlns aop: "http://www.springframework.org/schema/aop"
 
-            if (grailsApplication.config.org.transmart.security.samlEnabled) {
-                importBeans('classpath:/spring/spring-security-saml.xml')
-                // Provider of default SAML Context. Moved to groovy to allow choose implementation
-                if (grailsApplication.config.org.transmart.security.saml.lb.serverName) {
-                    contextProvider(org.springframework.security.saml.context.SAMLContextProviderLB) {
-                        scheme = grailsApplication.config.org.transmart.security.saml.lb.scheme
-                        serverName = grailsApplication.config.org.transmart.security.saml.lb.serverName
-                        serverPort = grailsApplication.config.org.transmart.security.saml.lb.serverPort
-                        includeServerPortInRequestURL = grailsApplication.config.org.transmart.security.saml.lb.includeServerPortInRequestURL
-                        contextPath = grailsApplication.config.org.transmart.security.saml.lb.contextPath
-                    }
-                } else {
-                    contextProvider(org.springframework.security.saml.context.SAMLContextProviderImpl)
-                }
-            }
-
             //overrides bean implementing GrailsUserDetailsService?
             userDetailsService(AuthUserDetailsService)
